Baikal-M: Add support for LVDS type vdu.
The patch detects (from device-tree) that the controller is lvds type,
determines number of active lanes and sets GPIOR appropriately.

Framebuffer start for vdu_hdmi is updated via smc call.
A similar call for vdu_lvds does not exist, yet.
So, we update fb start the same way as existed in sdk-4.1.