Baikal-M devicetree: Fix clocks, resolve warnings, cleanup
Some of the "clock-names" in device nodes have wrong names
(not that is expected by driver). This may result in
configuration errors or incorrect frequency detection.

The patch fixes the names for timer and gmac.

Clock specifications removed from device nodes if the
driver does not look for clocks. Fixed-clocks which are
not referenced have been removed.

dtc compiler/decompiler prints a lot of warnings. These
are also resolved.