clk: Add baikal fixed factor clock
This is intended to replace allwinner,sun4i-a10-pll3-2x-clk
which is not supported in linux 6.1.
The purpose is to add CLK_SET_RATE_PARENT flag to
fixed-factor-clock. This allows seting requested vdu-lvds
pixel clock to the desired frequency by changing pll clock
which is x7 to the pixel clock.